Update wscripts for Waf 1.6
We update the waf setup so that we can use the newest release. The
newer releases are simultaneously usable with either Python 2 or
Python 3.

wafmisc.py is a file copied from the waf distribution in
waflib/extras/misc.py. Its author is Thomas Nagy and the license is
the three-clause BSD license
http://www.opensource.org/licenses/bsd-license.php

For the wscripts to work in all parts, we require newer Waf than the
latest released version (Waf 1.6.1p1).